refactor: detail state store 분리
