+++ src/pages/ProductDetailPage.js
@@ -1,4 +1,3 @@
-import { getProducts, getProduct } from "../api/productApi.js";
 import { cartStore } from "../stores/cartStore.js";
 import { Footer } from "../components/Footer.js";
 import { ProductDetailHeader } from "../components/ProductDetailHeader.js";
@@ -7,155 +6,7 @@ import { ProductDetailInfo } from "../components/ProductDetailInfo.js";
 import { ProductQuantitySelector } from "../components/ProductQuantitySelector.js";
 import { ProductActionButtons } from "../components/ProductActionButtons.js";
 import { RelatedProducts } from "../components/RelatedProducts.js";
-import { isTestEnvironment } from "../utils/isTestEnvironment.js";
-
-const MAX_STOCK = Number.MAX_SAFE_INTEGER;
-
-// 상품 상세 상태 관리
-const productDetailStore = {
-  state: {
-    product: null,
-    relatedProducts: [],
-    quantity: 1,
-    isLoading: true,
-    relatedProductsLoading: false,
-    error: null,
-    reviewCount: 0,
-    rating: 0,
-  },
-
-  async fetchProductDetail(productId) {
-    this.state.isLoading = true;
-    this.state.error = null;
-    this.state.relatedProducts = []; // 관련 상품 초기화
-    this.state.quantity = 1; // 수량 초기화
-    this.state.reviewCount = 0;
-    this.state.rating = 0;
-
-    if (!isTestEnvironment()) {
-      this.render();
-    }
-
-    try {
-      // 개별 상품 API 사용
-      const product = await getProduct(productId);
-      this.state.product = product;
-      this.state.reviewCount = product?.reviewCount ?? 0;
-      this.state.rating = product?.rating ?? 0;
-
-      if (!this.state.product) {
-        throw new Error("상품을 찾을 수 없습니다.");
-      }
-
-      // 관련 상품 로딩
-      this.loadRelatedProducts();
-    } catch (err) {
-      console.error("Failed to fetch product detail:", err);
-      this.state.error = err;
-    } finally {
-      this.state.isLoading = false;
-      this.render();
-    }
-  },
-
-  async loadRelatedProducts() {
-    this.state.relatedProductsLoading = true;
-    this.render(); // 로딩 상태 표시
-
-    try {
-      // 관련 상품만 효율적으로 가져오기 (같은 카테고리 기준)
-      const data = await getProducts({
-        category2: this.state.product.category2,
-        limit: 20,
-      });
-
-      this.state.relatedProducts = data.products
-        .filter((p) => p.productId !== this.state.product.productId)
-        .slice(0, 19); // 19개만
-    } catch (err) {
-      console.error("Failed to load related products:", err);
-    } finally {
-      this.state.relatedProductsLoading = false;
-      this.render();
-    }
-  },
-
-  increaseQuantity() {
-    const maxStock = this.state.product?.stock;
-    this.state.quantity = Math.min(this.state.quantity + 1, maxStock || MAX_STOCK);
-
-    const quantityInput = document.querySelector("#quantity-input");
-    if (quantityInput) {
-      quantityInput.value = this.state.quantity;
-    }
-  },
-
-  decreaseQuantity() {
-    this.state.quantity = Math.max(this.state.quantity - 1, 1);
-
-    const quantityInput = document.querySelector("#quantity-input");
-    if (quantityInput) {
-      quantityInput.value = this.state.quantity;
-    }
-  },
-
-  updateQuantity(newQuantity) {
-    // 빈 문자열이나 유효하지 않은 값 처리
-    if (!newQuantity || newQuantity === "" || isNaN(newQuantity)) {
-      this.state.quantity = 1;
-      const quantityInput = document.querySelector("#quantity-input");
-      if (quantityInput) {
-        quantityInput.value = this.state.quantity;
-      }
-      return;
-    }
-
-    // 입력값을 숫자로 변환하고 범위 제한
-    const quantity = parseInt(newQuantity);
-    const maxStock = this.state.product?.stock;
-    const newQuantityValue = Math.max(1, Math.min(quantity, maxStock || MAX_STOCK));
-
-    // 값이 실제로 변경된 경우에만 업데이트
-    if (this.state.quantity !== newQuantityValue) {
-      this.state.quantity = newQuantityValue;
-
-      const quantityInput = document.querySelector("#quantity-input");
-      if (quantityInput) {
-        quantityInput.value = this.state.quantity;
-      }
-    }
-  },
-
-  addToCart() {
-    if (this.state.product) {
-      const productWithQuantity = {
-        ...this.state.product,
-        quantity: this.state.quantity,
-      };
-      cartStore.addToCart(productWithQuantity);
-    }
-  },
-
-  reset() {
-    this.state.product = null;
-    this.state.relatedProducts = [];
-    this.state.quantity = 1;
-    this.state.isLoading = true;
-    this.state.relatedProductsLoading = false;
-    this.state.error = null;
-  },
-
-  // 렌더링 콜백 (외부에서 주입)
-  render: null,
-};
-
-// 전역으로 노출 (HTML onchange/oninput에서 접근 가능하도록)
-window.productDetailStore = productDetailStore;
-
-// 테스트 환경에서 전역 함수로 노출
-if (isTestEnvironment()) {
-  window.__TEST_PRODUCT_DETAIL_STATE__ = productDetailStore;
-}
+import { productDetailStore } from "../stores/productDetailStore.js";
 
 function renderLoading() {
   return /* HTML */ `
@@ -227,16 +78,32 @@ export function initializeProductDetailPage() {
 }
 
 function setupEventListeners() {
-  // 이벤트 위임을 사용하여 document 레벨에서 이벤트 처리
-  document.addEventListener("click", (e) => {
+  // 기존 이벤트 리스너 제거 (중복 방지)
+  const existingListener = document._productDetailListener;
+  if (existingListener) {
+    document.removeEventListener("click", existingListener);
+  }
+
+  // 새로운 이벤트 리스너 생성
+  const clickListener = (e) => {
     // 수량 증가 버튼
     if (e.target.closest("#quantity-increase")) {
       productDetailStore.increaseQuantity();
+      // DOM 업데이트 확인
+      const quantityInput = document.querySelector("#quantity-input");
+      if (quantityInput) {
+        quantityInput.value = productDetailStore.state.quantity;
+      }
     }
 
     // 수량 감소 버튼
     if (e.target.closest("#quantity-decrease")) {
       productDetailStore.decreaseQuantity();
+      // DOM 업데이트 확인
+      const quantityInput = document.querySelector("#quantity-input");
+      if (quantityInput) {
+        quantityInput.value = productDetailStore.state.quantity;
+      }
     }
 
     // 장바구니 담기 버튼
@@ -262,7 +129,11 @@ function setupEventListeners() {
       window.history.pushState({}, "", `/${queryString}`);
       window.dispatchEvent(new Event("popstate"));
     }
-  });
+  };
+
+  // 이벤트 리스너 등록
+  document.addEventListener("click", clickListener);
+  document._productDetailListener = clickListener;
 
   // input의 value 관력 이벤트는 직접 바인딩 (이벤트 위임이 어려움)
   const quantityInput = document.querySelector("#quantity-input");
